+++ src/main/java/com/coffee_shop/coffeeshop/service/coupon/issue/fail/CouponIssueFailHandler.java
@@ -1,0 +1,7 @@
+package com.coffee_shop.coffeeshop.service.coupon.issue.fail;
+
+import com.coffee_shop.coffeeshop.service.coupon.dto.request.CouponApplication;
+
+public interface CouponIssueFailHandler {
+	void handleFail(CouponApplication couponApplication, Exception exception);
+}
