+++ src/test/java/com/coffee_shop/coffeeshop/service/coupon/issue/fail/CouponIssueFailHandlerImplTest.java
@@ -1,4 +1,4 @@
-package com.coffee_shop.coffeeshop.service.coupon.issue;
+package com.coffee_shop.coffeeshop.service.coupon.issue.fail;
 
 import static com.coffee_shop.coffeeshop.domain.coupon.CouponType.*;
 import static org.assertj.core.api.Assertions.*;
@@ -30,12 +30,13 @@ import com.coffee_shop.coffeeshop.service.IntegrationTestSupport;
 import com.coffee_shop.coffeeshop.service.coupon.apply.CouponApplyService;
 import com.coffee_shop.coffeeshop.service.coupon.dto.request.CouponApplication;
 import com.coffee_shop.coffeeshop.service.coupon.dto.request.CouponApplyServiceRequest;
+import com.coffee_shop.coffeeshop.service.coupon.issue.CouponIssueServiceImpl;
 
 import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.read.ListAppender;
 
-class CouponIssueFailHandlerTest extends IntegrationTestSupport {
+class CouponIssueFailHandlerImplTest extends IntegrationTestSupport {
 	@Autowired
 	private UserRepository userRepository;
 
@@ -43,7 +44,7 @@ class CouponIssueFailHandlerTest extends IntegrationTestSupport {
 	private CouponRepository couponRepository;
 
 	@Autowired
-	private CouponIssueFailHandler couponIssueFailHandler;
+	private CouponIssueFailHandlerImpl couponIssueFailHandlerImpl;
 
 	@Autowired
 	private CouponApplyService couponApplyService;
@@ -55,7 +56,7 @@ class CouponIssueFailHandlerTest extends IntegrationTestSupport {
 	private CouponTransactionHistoryRepository couponTransactionHistoryRepository;
 
 	@SpyBean
-	private CouponIssueService couponIssueService;
+	private CouponIssueServiceImpl couponIssueService;
 
 	private Long exceptionUserId;
 
@@ -78,7 +79,7 @@ class CouponIssueFailHandlerTest extends IntegrationTestSupport {
 		doThrow(new RuntimeException()).when(couponIssueService).issueCoupon(any(CouponApplication.class));
 
 		ListAppender<ILoggingEvent> listAppender = new ListAppender<>();
-		Logger logger = (Logger)LoggerFactory.getLogger(CouponIssueFailHandler.class);
+		Logger logger = (Logger)LoggerFactory.getLogger(CouponIssueFailHandlerImpl.class);
 		logger.addAppender(listAppender);
 		listAppender.start();
 
@@ -118,7 +119,7 @@ class CouponIssueFailHandlerTest extends IntegrationTestSupport {
 
 		//log 체크
 		ListAppender<ILoggingEvent> listAppender = new ListAppender<>();
-		Logger logger = (Logger)LoggerFactory.getLogger(CouponIssueFailHandler.class);
+		Logger logger = (Logger)LoggerFactory.getLogger(CouponIssueFailHandlerImpl.class);
 		logger.addAppender(listAppender);
 		listAppender.start();
 
@@ -185,7 +186,7 @@ class CouponIssueFailHandlerTest extends IntegrationTestSupport {
 
 		//log 체크
 		ListAppender<ILoggingEvent> listAppender = new ListAppender<>();
-		Logger logger = (Logger)LoggerFactory.getLogger(CouponIssueFailHandler.class);
+		Logger logger = (Logger)LoggerFactory.getLogger(CouponIssueFailHandlerImpl.class);
 		logger.addAppender(listAppender);
 		listAppender.start();
